demos/researcher_agent/requirements.txt
streamlit>=1.28.0
agno>=0.1.0
memorisdk>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
litellm>=1.0.0
httpx>=0.25.0
exa_py==1.14.20